agent-ros-bridge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-ros-bridge"
version = "0.4.0"
description = "Secure, multi-agent bridge for ROS1/2 robot control"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
authors = [
    {name = "OpenClaw", email = "hello@openclaw.ai"}
]
keywords = ["ros", "robotics", "mcp", "ai", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: Robot Framework",
]

dependencies = [
    # Core dependencies (minimal)
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
]
# ROS1 support requires system ROS1 installation (no pip package)
# Install from: http://wiki.ros.org/noetic/Installation
ros1 = []
ros2 = [
    "rclpy>=3.0.0",
]
websocket = [
    "websockets>=10.0",
]
grpc = [
    "grpcio>=1.50.0",
]
openclaw = [
    "aiohttp>=3.8.0",
]
yaml = [
    "pyyaml>=6.0",
]

# Phase 3: Observability
metrics = [
    "prometheus-client>=0.15.0",
]
tracing = [
    "opentelemetry-api>=1.15.0",
    "opentelemetry-sdk>=1.15.0",
    "opentelemetry-exporter-jaeger>=1.15.0",
]
dashboard = [
    "aiohttp>=3.8.0",
]

all = [
    "mcp>=1.0.0",
    "rclpy>=3.0.0",
    "websockets>=10.0",
    "grpcio>=1.50.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.15.0",
    "opentelemetry-api>=1.15.0",
    "opentelemetry-sdk>=1.15.0",
    "redis>=4.5.0",
]

dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0",
    "mypy>=1.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/openclaw/agent-ros-bridge"
Documentation = "https://docs.openclaw.ai/agent-ros-bridge"
Repository = "https://github.com/openclaw/agent-ros-bridge"
Issues = "https://github.com/openclaw/agent-ros-bridge/issues"
Changelog = "https://github.com/openclaw/agent-ros-bridge/blob/main/CHANGELOG.md"
"PyPI" = "https://pypi.org/project/agent-ros-bridge/"
"ClawHub" = "https://clawhub.ai/skills/agent-ros-bridge"
"OpenClaw" = "https://openclaw.ai"

[project.scripts]
agent-ros-bridge = "agent_ros_bridge.mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["agent_ros_bridge"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

# Phase 1: Testing Configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (may use external services)",
    "e2e: End-to-end tests (full system)",
    "slow: Slow tests",
    "requires_ros: Tests requiring ROS installation",
    "requires_mcp: Tests requiring MCP package",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["agent_ros_bridge"]
omit = [
    "*/tests/*",
    "*/test_*",
    "agent_ros_bridge/mcp/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
fail_under = 80
